refactor: remove pacotes antigos e organiza estrutura
