--- systemvigiasus/src/main/java/com/systemvigiasus/backend/SystemvigiasusApplication.java
@@ -1,13 +1,0 @@
-package com.systemvigiasus.backend;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SystemvigiasusApplication {
-
-	public static void main(String[] args) {
-		SpringApplication.run(SystemvigiasusApplication.class, args);
-	}
-
-}
