--- systemvigiasus/src/main/java/com/systemvigiasus/backend/TesteController.java
@@ -1,13 +1,0 @@
-package com.systemvigiasus.backend;
-
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class TesteController {
-
-    @GetMapping("/teste")
-    public String teste() {
-        return "API funcionando";
-    }
-}
