+++ systemvigiasus/src/main/java/com/systemvigiasus/monitoramento/SystemvigiasusApplication.java
@@ -1,0 +1,13 @@
+package com.systemvigiasus.monitoramento;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SystemvigiasusApplication {
+
+	public static void main(String[] args) {
+		SpringApplication.run(SystemvigiasusApplication.class, args);
+	}
+
+}
