--- systemvigiasus/src/test/java/com/systemvigiasus/backend/SystemvigiasusApplicationTests.java
@@ -1,13 +1,0 @@
-package com.systemvigiasus.backend;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class SystemvigiasusApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
