+++ systemvigiasus/src/test/java/com/systemvigiasus/monitoramento/SystemvigiasusApplicationTests.java
@@ -1,0 +1,13 @@
+package com.systemvigiasus.monitoramento;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SystemvigiasusApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
